requirements.txt
torch
torchvision
pillow
pandas
tqdm
transformers>=4.56.0
accelerate